requirements.txt
Quart
qdrant-client
sentence-transformers
torch
numpy
asyncio
python-dotenv